+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - (Feature) (Platform) Add Sharding to the Inventory Endpoint
 - (Feature) Emit Default values in Inventory
 - (Maintenance) Bump github.com/containerd/containerd to v1.7.27
+- (Feature) Rotation Order
 
 ## [1.2.46](https://github.com/arangodb/kube-arangodb/tree/1.2.46) (2025-02-24)
 - (Bugfix) Clean Phase change properly during upgrade